+++ R/sensors-mspct.r
@@ -1,7 +1,7 @@
 #' @title Spectral response of sensors
 #'
 #' @description A collection of response spectra for various broadband sensors
-#' used for measuring ultraviolet and visible radiation. Each spectrun in the
+#' used for measuring ultraviolet and visible radiation. Each spectrum in the
 #' collection contains two variables, wavelengths (nm) at either regular or
 #' irregular intervals and spectral responsiveness (in energy units).
 #' Spectral data are in most cases normalized to one at the wavelength of
@@ -11,7 +11,7 @@
 #' @format A \code{response_mspct} object containing
 #' \code{response_spct} objects as \emph{named} members.
 #'
-#' Spectral objects for single channel sensers contain two numeric variables,
+#' Spectral objects for single channel sensors contain two numeric variables,
 #' with responsivity in most cases in relative energy units:
 #' \itemize{
 #'    \item w.length (nm)
@@ -28,7 +28,7 @@
 #' stored in R attributes.
 #'
 #' @note In addition to this object containing the spectral data, this package
-#' provides character vectors useful for subsetting spectra by supplier, type,
+#' provides character vectors useful for sub-setting spectra by supplier, type,
 #' colour, etc.
 #'
 #' @docType data
@@ -41,3 +41,43 @@
 #' names(sensors.mspct)
 #'
 "sensors.mspct"
+
+#' @title Spectral response of image sensors
+#'
+#' @description A collection of response spectra for various image sensors and
+#' digital cameras responsive to ultraviolet, visible, NIR and/or SWIR
+#' radiation. Each spectrum in the collection contains two variables,
+#' wavelengths (nm) at irregular intervals and spectral responsiveness (in
+#' energy units).
+#' Spectral data are in most cases not normalized to one at the wavelength of
+#' maximum energy responsivity.
+#'
+#' @format A \code{response_mspct} object containing
+#' \code{response_spct} objects as \emph{named} members.
+#'
+#' Spectral objects for monochrome image sensors contain two numeric variables,
+#' with responsivity in most cases in relative energy units:
+#' \itemize{
+#'    \item w.length (nm)
+#'    \item s.e.response (r.u.)
+#' }
+#' Spectral objects for colour image sensors contain the spectra in long form
+#' with two numeric variables, and in addition a factor with channel ids:
+#' \itemize{
+#'    \item w.length (nm)
+#'    \item s.e.response (r.u.)
+#'    \item channel (factor with names as levels)
+#' }
+#' The spectral objects contain, in addition to the spectral data, metadata
+#' stored in R attributes.
+#'
+#' @docType data
+#' @keywords datasets
+#'
+#' @seealso \code{\link[photobiology:source_spct]{response_spct}} and
+#' \code{\link[photobiology:generic_mspct]{response_mspct}}
+#'
+#' @examples
+#' names(image_sensors.mspct)
+#'
+"image_sensors.mspct"